perf: switch to modppl's f32 feature, dropping the f64/f32 cast dance
modppl 0.3.1 adds an f32 feature (Real = f32 instead of f64), matching the
renderer's native precision (glam is f32 throughout). Removes ~150 "as f32"
/ "as f64" conversions across models, proposals, and the pose-net CNN
pipeline; the few that remain are genuine boundary crossings (float_extras'
erf has no f32 variant, so the truncated-normal CDF stays f64 internally).

Also fixes the runtime-typed half of the migration that the compiler can't
catch: every trace.data.read::<f64>() and Arc::new(..._f64) observation
constraint, which would otherwise panic on a downcast mismatch rather than
fail to build.

Measured ~23% faster on the mug inference benchmark (6.5ms/iter -> 5.0ms/iter),
apparently from the per-pixel truncated-normal scoring loop no longer
round-tripping through f64.

+++ derender/src/baseline/ray.rs
@@ -12,13 +12,13 @@ struct UniformS2 { }
 const uniform_s2: UniformS2 = UniformS2 { };
 
 impl Distribution<Vec3,()> for UniformS2 {
-    fn logpdf(&self, _: &Vec3, _: ()) -> f64 {
+    fn logpdf(&self, _: &Vec3, _: ()) -> f32 {
         panic!("not implemented!");
     }
-    
+
     fn random(&self, rng: &mut ThreadRng, _: ()) -> Vec3 {
-        let z = 2.0 * u01(rng) as f32 - 1.0;
-        let theta = 2.0 * PI * u01(rng) as f32;
+        let z = 2.0 * u01(rng) - 1.0;
+        let theta = 2.0 * PI * u01(rng);
         let r = (1.0 - z * z).sqrt();
         [r * theta.cos(), r * theta.sin(), z]
     }
