+++ derender/src/core/proposals.rs
@@ -3,10 +3,10 @@ use crate::image::*;
 
 
 dyngen!(
-pub fn rgbd_drift(trace: Weak<DynTrace<(f32,f32,bool),(Depths,Colors)>>, mask: Vec<&str>, stdev: f64) {
+pub fn rgbd_drift(trace: Weak<DynTrace<(f32,f32,bool),(Depths,Colors)>>, mask: Vec<&str>, stdev: f32) {
     let trace = trace.upgrade().unwrap();
     for addr in mask.iter() {
-        normal(trace.data.read::<f64>(addr), stdev) %= addr;
+        normal(trace.data.read::<f32>(addr), stdev) %= addr;
     }
 });
 
@@ -20,24 +20,24 @@ dyngen!(
 /// `targets`: (address, center, stdev) triples. Out-of-support proposals
 /// (e.g. azimuth outside [0, 2π)) score -inf under the model prior and are
 /// simply rejected -- same behavior as the existing drift moves near bounds.
-pub fn pose_guide(_trace: Weak<DynTrace<(f32,f32,bool),(Depths,Colors)>>, targets: Vec<(&str, f64, f64)>) {
+pub fn pose_guide(_trace: Weak<DynTrace<(f32,f32,bool),(Depths,Colors)>>, targets: Vec<(&str, f32, f32)>) {
     for (addr, center, stdev) in targets.iter() {
         normal(*center, *stdev) %= addr;
     }
 });
 
 dyngen!(
-pub fn gaussian_drift(trace: Weak<DynTrace<(),Colors>>, mask: Vec<&str>, stdev: f64) {
+pub fn gaussian_drift(trace: Weak<DynTrace<(),Colors>>, mask: Vec<&str>, stdev: f32) {
     let trace = trace.upgrade().unwrap();
     for addr in mask.iter() {
-        normal(trace.data.read::<f64>(addr), stdev) %= addr;
+        normal(trace.data.read::<f32>(addr), stdev) %= addr;
     }
 });
 
 dyngen!(
-pub fn noise_drift(trace: Weak<DynTrace<f32,Colors>>, mask: Vec<&str>, stdev: f64) {
+pub fn noise_drift(trace: Weak<DynTrace<f32,Colors>>, mask: Vec<&str>, stdev: f32) {
     let trace = trace.upgrade().unwrap();
     for addr in mask.iter() {
-        normal(trace.data.read::<f64>(addr), stdev) %= addr;
+        normal(trace.data.read::<f32>(addr), stdev) %= addr;
     }
 });
